Update README.md to reflect v2 incompatibility
Hi @sampsyo! 

Thanks for the work on the SmartThinq API. The library doesn't seem to currently work on the v2 API, and until it does, I suggest this small change to the readme so that people aren't creating issues or are confused as to why it doesn't work for them. I was confused for a little bit until I went through the issue tracker to figure out why I wasn't getting any devices.

Of course, happy to help get v2 on track.